unit3-storage/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
python-dotenv==1.0.0
httpx==0.25.1
firebase-admin==6.2.0
cloudinary==1.36.0
